+++ posthog-android/src/main/java/com/posthog/android/replay/PostHogReplayIntegration.kt
@@ -33,6 +33,8 @@ import android.widget.RatingBar
 import android.widget.Spinner
 import android.widget.Switch
 import android.widget.TextView
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import com.posthog.PostHogIntegration
 import com.posthog.android.PostHogAndroidConfig
 import com.posthog.android.internal.MainHandler
@@ -44,6 +46,7 @@ import com.posthog.android.replay.internal.PostHogLogCatWatcher
 import com.posthog.android.replay.internal.ViewTreeSnapshotStatus
 import com.posthog.internal.PostHogThreadFactory
 import com.posthog.internal.replay.RRAddedNode
+import com.posthog.internal.replay.RRCustomEvent
 import com.posthog.internal.replay.RREvent
 import com.posthog.internal.replay.RRFullSnapshotEvent
 import com.posthog.internal.replay.RRIncrementalMouseInteractionData
@@ -106,6 +109,34 @@ public class PostHogReplayIntegration(
                             executor.submit {
                                 generateSnapshot(WeakReference(decorView), timestamp)
                             }
+                        }
+
+                        // keyboard events
+                        var open = false
+                        ViewCompat.setOnApplyWindowInsetsListener(view) { _, insets ->
+                            val imeVisible = insets.isVisible(WindowInsetsCompat.Type.ime())
+                            if (open == imeVisible) {
+                                return@setOnApplyWindowInsetsListener insets
+                            }
+                            open = imeVisible
+
+                            val payload = mutableMapOf<String, Any>()
+                            if (imeVisible) {
+                                val imeHeight = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
+                                payload["open"] = true
+                                payload["height"] = imeHeight.densityValue(displayMetrics.density)
+                            } else {
+                                payload["open"] = false
+                            }
+
+                            val event = RRCustomEvent(
+                                tag = "keyboard",
+                                payload = payload,
+                                config.dateProvider.currentTimeMillis(),
+                            )
+                            listOf(event).capture()
+
+                            insets
                         }
 
                         val status = ViewTreeSnapshotStatus(listener)