+++ posthog/src/main/java/com/posthog/internal/PostHogQueue.kt
@@ -87,7 +87,7 @@ internal class PostHogQueue(
                     }
                     config.logger.log("Queued event ${file.name}.")
 
-                    flushIfOverThreshold(true)
+                    flushIfOverThreshold()
                 } catch (e: Throwable) {
                     config.logger.log("Event ${event.event} failed to parse: $e.")
                 }
@@ -95,10 +95,14 @@ internal class PostHogQueue(
         }
     }
 
-    private fun flushIfOverThreshold(onExecutor: Boolean) {
-        if (deque.size >= config.flushAt) {
-            flushBatch(onExecutor)
+    private fun flushIfOverThreshold() {
+        if (isAboveThreshold(config.flushAt)) {
+            flushBatch()
         }
+    }
+
+    private fun isAboveThreshold(flushAt: Int): Boolean {
+        return deque.size >= flushAt
     }
 
     private fun canFlushBatch(): Boolean {
@@ -118,7 +122,7 @@ internal class PostHogQueue(
         return events
     }
 
-    private fun flushBatch(onExecutor: Boolean) {
+    private fun flushBatch() {
         if (!canFlushBatch()) {
             config.logger.log("Cannot flush the Queue.")
             return
@@ -129,14 +133,7 @@ internal class PostHogQueue(
             return
         }
 
-        // if its called from the executor already, no need to schedule another one
-        if (onExecutor) {
-            executeBatch()
-        } else {
-            executor.executeSafely {
-                executeBatch()
-            }
-        }
+        executeBatch()
     }
 
     private fun executeBatch() {
@@ -217,8 +214,8 @@ internal class PostHogQueue(
     }
 
     fun flush() {
-        if (!canFlushBatch()) {
-            config.logger.log("Cannot flush the Queue.")
+        // only flushes if the queue is above the threshold (not empty in this case)
+        if (!isAboveThreshold(1)) {
             return
         }
 
@@ -274,7 +271,8 @@ internal class PostHogQueue(
                     config.logger.log("Queue is flushing.")
                     return@schedule
                 }
-                flushIfOverThreshold(false)
+                // if the timer passes, send pending events, no need to wait for the flushAt threshold
+                flush()
             }
             this.timerTask = timerTask
             this.timer = timer